Moved tile_location_helpers
