+++ lib/src/widgets/game_board.dart
@@ -4,7 +4,7 @@ import 'package:link_five/src/logic/actions/place_tile_action.dart';
 import 'package:link_five/src/model/game/game_state.dart';
 import 'package:link_five/src/model/game/player_color.dart';
 import 'package:link_five/src/model/game/tile_location.dart';
-import 'package:link_five/src/model/game/tile_location_helpers.dart';
+import 'package:link_five/src/widgets/tile_location_helpers.dart';
 import 'package:link_five/src/widgets/game_board/constants.dart';
 import 'package:link_five/src/widgets/game_board/tile_painter.dart';
